% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X.kc}
\alias{X.kc}
\title{Example Inspection Scores Matrix.}
\format{A matrix with 4 columns and ~1500 rows, where each row represents
  one business and each column is one inspection cycle.
  \code{X.kc[i,j]} represents the inspection score for the
  \code{i}th restaurant in the \code{j}th most recent inspection.}
\usage{
X.kc
}
\description{
A small dataset of inspection scores.
}
\details{
\code{X.kc} contains restaurant inspection information from 11 randomly
chosen ZIP codes in the King County (WA) jurisdiction. Establishments and ZIP
codes are masked. Inspection information is limited to the 01-01-2012 to
03-25-2016 time period.
}
\keyword{datasets}

